emotional_crystal_module/requirements.txt

streamlit
pandas
numpy
requests
Pillow
nltk
